chore: Refactor interceptors to Interceptors namespace
Moved activity and metrics interceptors and their tests from .Internals to .Interceptors namespace for improved clarity. Updated static imports and references in configuration and test files to match the new organization. This enhances code maintainability and separation of concerns.
