+++ tests/NetEvolve.Pulse.Tests.Unit/ServiceCollectionExtensionsTests.cs
@@ -3,6 +3,7 @@ namespace NetEvolve.Pulse.Tests.Unit;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using NetEvolve.Pulse.Extensibility;
+using NetEvolve.Pulse.Interceptors;
 using NetEvolve.Pulse.Internals;
 using TUnit.Core;
